% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coman.R
\name{coman}
\alias{coman}
\title{A function that does a complte DEGA from start to finish}
\usage{
coman(mydir, group, design, contrast)
}
\arguments{
\item{mydir}{The directory that contains your data.}

\item{group}{Group of factors in your experiment}

\item{design}{Design matrix with named columns}

\item{contrast}{contrast matrix}
}
\description{
This function allows you to do a complete differntial gene expression
analysis your hugene 2.0 St array data.
It will return three global variables. raWData which is the unnormalized gene expression values and
normData which are the log2 normalized gene expression values. Results which is the results of the analysis.
}
\examples{
group <- factor(c((rep.int(0,13)),rep.int(1,14),rep.int(2,14)
rep.int(3,14),rep.int(4,14)))
design <- model.matrix(~ 0 + group)
colnames(design) <- c("I","T1","T2","T3", "T4")
contrast matrix 
contrast <- makeContrasts( "I-T1","T1-T2","T2-T3","T3-T4",
                          levels= design )

coman(mydir, group, design, contrast)
}
\keyword{hugene,}
\keyword{microarrays}

