% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter}
\alias{filter}
\title{A Function to filter genes}
\usage{
filter(normData)
}
\arguments{
\item{normData}{The normalized data.}
}
\description{
This function will filter out genes that don't change throughout the time points
it will also filter out the coontrol probs
}
\examples{
filter()
}
\keyword{hugene,}
\keyword{microarrays}

