% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filterc}
\alias{filterc}
\title{A Function to filter genes}
\usage{
filterc(normData)
}
\arguments{
\item{normData}{The normalized data.}
}
\description{
This function will filter out genes that don't change throughout the time points
it will also filter out the control probs. It will return the number of features
filtered.
}
\examples{
filterc()
}
\keyword{hugene,}
\keyword{microarrays}

