% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lonoan.R
\name{lonoan}
\alias{lonoan}
\title{A Function to Load, Normalize, and Annotate}
\usage{
lonoan(mydir)
}
\arguments{
\item{mydir}{The directory that contains your data.}
}
\description{
This function allows you load, normalize, and annotate your hugene 2.0 St array data.
It will return two global variables. raWData which is the unnormalized gene expression values and
normData which are the log2 normalized gene expression values.
}
\examples{
lonoan()
}
\keyword{hugene,}
\keyword{microarrays}

