% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{qc}
\alias{qc}
\title{A quality control function}
\usage{
qc(rawData, normData)
}
\arguments{
\item{rawData}{The raw data. normData The normalized data.}
}
\description{
This function will allow you to see the boxplots of raw and normal data as well as 
a cluster analysis of the normalized data.
}
\examples{
qc()
}
\keyword{hugene,}
\keyword{microarrays}

