% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normfit.R
\name{normfit}
\alias{normfit}
\title{A function that returns your DEG in the form of nromfit}
\usage{
normfit(normData)
}
\arguments{
\item{normData}{Normalized data.}
}
\description{
Then it will fit a linear models to determine the mean values of your genes
and use ebayes to shrink the variance
}
\examples{
normfit()
}
\keyword{hugene,}
\keyword{microarrays}

